iox-#19-add-relocatable-pointers

+++ iceoryx_posh/include/iceoryx_posh/internal/popo/receiver_handler.inl
@@ -226,7 +226,7 @@ inline uint32_t ReceiverHandler<MaxReceivers, LockingPolicy>::getMaxDeliveryFiFo
         }
     }
 
-    return maxDeliveryFiFoCapacity;
+    return static_cast<uint32_t>(maxDeliveryFiFoCapacity);
 }
 
 } // namespace popo
